Add files via upload
These are some projects which I created while learning IOS Development:

BMI Calculator: This is a basic BMI calculator which will tell whether your BMI is good or not and will recommend some things that whether you should eat some more pies or not.

Destini: In this app, you will be creating a “choose your own adventure” game similar to the App Store hit “Life Line” app. The app will tell a story depending on what the user chooses and can be fleshed out and modified to provide an engaging story-telling experience

EggTimer: Build a beautiful egg timer app to boil your eggs to perfection depending on how you prefer your eggs. 

Quizzler: It is a trivia quiz app. If you have another field of expertise such as law or medicine, you’ll be glad to know that multiple-choice question apps are one of the most popular types of educational apps on the App Store! 

Xylophone:  This is my first musical instrument! Music apps are so popular on the App Store that they even get their own category. So in this module, we’re going to make a colorful XyloPhone app.